refactor(fe): update API request url
